rename modal component from vlan to ipv6_prefix
