+++ packages/frinx-uniresource-ui/src/pages/pool-detail-page/claim-resource-modal/claim-resource-allocating-modals/claim-resource-allocating-ipv6_prefix.tsx
